[sc-94] create food call API to create food, navigate to log created food after the API call, disable create food form while calling the API

+++ lib/features/add_food/add_food_controller.dart
@@ -42,7 +42,7 @@ class AddFoodController {
   int get proteinPercentage => _nutrition.proteinPercentage;
 
   void recalculateNutrition({required String servingSizeGrams}) {
-    final serving = int.tryParse(servingSizeGrams) ?? 100;
+    final serving = double.tryParse(servingSizeGrams) ?? 100;
     currentServingSizeNutrients.value = Nutrition.perServing(nutritionPer100Grams: _nutrition, servingSizeGrams: serving);
   }
